Squashed commit of the following:
commit dd54657d48866f62e488ad5747402aecaec72419
Author: juankisuelamartin <144780658+juankisuelamartin@users.noreply.github.com>
Date:   Wed Nov 1 13:38:16 2023 +0100

    Añadido SySos

    Añadimos System.out's para tener localizado donde añadir ventanas emergentes en el producto final.


commit c0eb8bdfe4ee979888bfd721b781bb4fb91912e4
Author: juankisuelamartin <144780658+juankisuelamartin@users.noreply.github.com>
Date:   Wed Nov 1 13:32:36 2023 +0100

    F-GESTION PRESTAMOS FINALIZADO

    INTERFAZ BETA FINALIZADA.
    Siguiente Sprint se realizaran pruebas de todas las funcionalidades implementadas para buscar errores.


commit 3f235bb9f56cd4d02ceb4ca00f545fbd575b2215
Author: juankisuelamartin <144780658+juankisuelamartin@users.noreply.github.com>
Date:   Wed Nov 1 12:04:16 2023 +0100

    REALIZAR DEVOLUCION

    Completed


commit 10e08938fba69e1e0b07def82cd89355080bd83d
Author: juankisuelamartin <144780658+juankisuelamartin@users.noreply.github.com>
Date:   Wed Nov 1 11:54:42 2023 +0100

    ALTA PRESTAMO FINALIZADO


commit 548873638ee64e1fe92c1d2c2f55bbb142a41f9e
Author: juankisuelamartin <144780658+juankisuelamartin@users.noreply.github.com>
Date:   Wed Nov 1 11:23:50 2023 +0100

    FUNCIONA ALTAPRESTAMO

commit 41e969b4066849133628b10c2b35c8c0e55adf67
Author: juankisuelamartin <144780658+juankisuelamartin@users.noreply.github.com>
Date:   Tue Oct 31 19:47:00 2023 +0100

    GestionPrestamos

    Work in progress


commit f63df503de2d5c325f595fffde96909df866f4de
Author: juankisuelamartin <144780658+juankisuelamartin@users.noreply.github.com>
Date:   Tue Oct 31 12:26:31 2023 +0100

    AltaPrestamo

    Error. Arreglar


commit 26a8875093223957f4d24676ec8dcb5ddc2f846b
Author: juankisuelamartin <144780658+juankisuelamartin@users.noreply.github.com>
Date:   Mon Oct 30 19:44:15 2023 +0100

    AltaPrestamo

+++ src/main/java/PantysMelRep/domain/controllers/GestorTitulos.java
@@ -1,10 +1,7 @@
 package PantysMelRep.domain.controllers;
 
 import PantysMelRep.domain.entities.*;
-import PantysMelRep.persistencia.AgenteBBDD;
-import PantysMelRep.persistencia.AutorDAO;
-import PantysMelRep.persistencia.EjemplarDAO;
-import PantysMelRep.persistencia.TituloDAO;
+import PantysMelRep.persistencia.*;
 
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -13,6 +10,7 @@ import org.springframework.stereotype.Service;
 
 
 import java.util.Collection;
+import java.util.Optional;
 
 @Service
 public class GestorTitulos {
@@ -23,6 +21,8 @@ public class GestorTitulos {
 	EjemplarDAO ejemplarDAO;
 	@Autowired
 	AutorDAO autorDAO;
+	@Autowired
+	PrestamoDAO prestamoDAO;
 	private AgenteBBDD agente;
 
 
@@ -73,7 +73,6 @@ public class GestorTitulos {
 	public void borrarTitulo(String isbn) {
 		Titulo titulo = tituloDAO.findById(isbn).orElseThrow(() -> new RuntimeException("Título no encontrado"));
 		tituloDAO.delete(titulo);
-		System.out.println("Titulo y ejemplares borrados");
 	}
 	@Transactional
 	public void altaEjemplar(String isbn) {
@@ -114,14 +113,30 @@ public class GestorTitulos {
 		// Buscar el ejemplar en la base de datos
 		Ejemplar ejemplar = ejemplarDAO.findById(id).orElseThrow(() -> new RuntimeException("Ejemplar no encontrado"));
 		if (ejemplar != null) {
-			ejemplarDAO.delete(ejemplar);
-			System.out.println("Ejemplar borrado");
+			System.out.println(Long.parseLong(id));
+			if(prestamoDAO.findByejemplarId(Long.parseLong(id)).isPresent()){
+				Prestamo prestamo = prestamoDAO.findByejemplarId(Long.parseLong(id)).orElseThrow(() -> new RuntimeException("Prestamo no encontrado"));
+				if(prestamo!= null){
+					if (prestamo.isActivo()){
+						System.out.println("El ejemplar no se puede borrar porque está prestado");
+					}
+					else{
+						prestamoDAO.delete(prestamo);
+						ejemplarDAO.delete(ejemplar);
+						System.out.println("Ejemplar borrado. y prestamo inactivo borrado.");
+					}
+			}
+			}else{
+				ejemplarDAO.delete(ejemplar);
+				System.out.println("Ejemplar borrado.");
+
+
 		}
 	}
 
 
 
-
+}
 
 
 }
